[feature#162723781] added validator.py file for validating username, password and email fields
