+++ app/api/v1/users/models.py
@@ -1,4 +1,4 @@
-from app.api.v1.users.utils import validate_password, validate_username, validate_email
+from app.api.v1.users.validators import validate_password, validate_username, validate_email
 from app import create_app, bcrypt
 
 users_list = [
